Server/requirements.txt
aniso8601==9.0.1
attrs
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.1
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
packaging==22.0
pkgutil-resolve-name
pluggy==1.0.0
py==1.11.0
PyJWT==2.6.0
pyrsistent==0.19.2
pytest==7.1.2
python-dotenv==0.19.2
pytz==2022.6
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.46
tomli==2.0.1
urllib3==1.26.13
Werkzeug==2.1.2
zipp==3.11.0
flask_mysqldb
pymysql
# psycopg2-binary
